requirements.txt
transformers==4.30.2
torch==2.0.1
Pillow==9.5.0
bitsandbytes==0.43.3
python-dotenv
flask